+++ src/routes/(v1)/multiple/+page.server.ts
@@ -4,7 +4,8 @@ import { zod } from '$lib/adapters/index.js';
 import { error, fail } from '@sveltejs/kit';
 import type { Actions, PageServerLoad } from './$types.js';
 
-import { userSchema, users } from '../users';
+import { userSchema, users } from '../users.js';
+import { z } from 'zod';
 
 type Message = { message: string };
 
@@ -15,7 +16,7 @@ export const load = (async ({ url }) => {
 
 	if (id && !user) throw error(404, 'User not found.');
 
-	const first = await superValidate<typeof userSchema, Message>(user, userSchema);
+	const first = await superValidate<z.infer<typeof userSchema>, Message>(user, zod(userSchema));
 	const second = structuredClone(first);
 	second.data.name += ' the 2:nd';
 
@@ -29,7 +30,7 @@ export const actions = {
 
 		console.log('POST', data);
 
-		const posted = await superValidate<typeof userSchema, Message>(data, userSchema);
+		const posted = await superValidate<z.infer<typeof userSchema>, Message>(data, zod(userSchema));
 
 		console.log('FORM ', posted);
 		if (!posted.valid) return fail(400, { form: posted });
@@ -42,7 +43,7 @@ export const actions = {
 		// happens when multiple id's are posted.
 		if (posted.id == 'second') {
 			posted.data.name = '2:nd ' + posted.data.name;
-			other.id = (await superValidate(null, userSchema)).id;
+			other.id = (await superValidate(null, zod(userSchema))).id;
 		} else {
 			other.data.name = '2:nd ' + other.data.name;
 			other.id = 'second';