+++ src/routes/(v1)/nested-validation/TagForm.svelte
@@ -7,7 +7,6 @@
 	import * as flashModule from 'sveltekit-flash-message/client';
 	import { onMount } from 'svelte';
 	import type { SuperValidated } from '$lib/index.js';
-	import { zod } from '$lib/adapters/index.js';
 
 	import type { z } from 'zod';
 	import { zod } from '$lib/adapters/zod.js';